pyproject.toml
[project]
name = "pdf-ocr-mvp"
version = "0.1.0"
description = "Extracts text and token-level OCR from PDFs."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.110.0",
    "uvicorn[standard]==0.27.1",
    "python-multipart==0.0.9",
    "pytesseract==0.3.10",
    "pillow==10.2.0",
    "pdf2image==1.17.0",
    "tika==2.6.0",
    "pydantic==2.6.1",
    "opencv-python==4.11.0.86",
    "pymupdf>=1.24.0",
]
[project.optional-dependencies]
diagrams = [
    "openai>=1.12.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
